added security

+++ src/main/java/com/example/danceSchool/service/impl/AdminServiceImpl.java
@@ -4,11 +4,15 @@ import com.example.danceSchool.dto.AdminDto;
 import com.example.danceSchool.dto.LessonDto;
 import com.example.danceSchool.dto.SheduleReport;
 import com.example.danceSchool.dto.SheduleRowMapper;
+import com.example.danceSchool.dto.TeacherDto;
 import com.example.danceSchool.entity.Admin;
 import com.example.danceSchool.entity.Lesson;
+import com.example.danceSchool.entity.Teacher;
+import com.example.danceSchool.exception.AdminException;
 import com.example.danceSchool.exception.TeacherException;
 import com.example.danceSchool.repository.AdminRepository;
 import com.example.danceSchool.repository.LessonRepository;
+import com.example.danceSchool.repository.RoleRepository;
 import com.example.danceSchool.repository.TeacherRepository;
 import com.example.danceSchool.service.AdminService;
 import com.example.danceSchool.service.EmailService;
@@ -33,15 +37,17 @@ public class AdminServiceImpl implements AdminService {
     private final EmailService emailService;
     private final LessonRepository lessonRepository;
     private final TeacherRepository teacherRepository;
+    private final RoleRepository roleRepository;
 
     @Autowired
-    public AdminServiceImpl(AdminRepository adminRepository, ConversionService conversionService, JdbcTemplate jdbcTemplate, EmailService emailService, LessonRepository lessonRepository, TeacherRepository teacherRepository) {
+    public AdminServiceImpl(AdminRepository adminRepository, ConversionService conversionService, JdbcTemplate jdbcTemplate, EmailService emailService, LessonRepository lessonRepository, TeacherRepository teacherRepository, RoleRepository roleRepository) {
         this.adminRepository = adminRepository;
         this.conversionService = conversionService;
         this.jdbcTemplate = jdbcTemplate;
         this.emailService = emailService;
         this.lessonRepository = lessonRepository;
         this.teacherRepository = teacherRepository;
+        this.roleRepository = roleRepository;
     }
 
     @Override
@@ -53,6 +59,7 @@ public class AdminServiceImpl implements AdminService {
     @Override
     public AdminDto createAdmin(AdminDto adminDto) {
         Admin admin = conversionService.convert(adminDto, Admin.class);
+        admin.setRole(roleRepository.findByName("ROLE_ADMIN"));
         return conversionService.convert(adminRepository.save(admin), AdminDto.class);
     }
 
@@ -73,9 +80,12 @@ public class AdminServiceImpl implements AdminService {
 
     @Override
     public void deleteAdmin(Long id) {
-        Admin admin = adminRepository.findById(id).orElseThrow(() -> new TeacherException("Admin is not found"));
-        adminRepository.delete(admin);
-
+        if (adminRepository.count() > 1) {
+            Admin admin = adminRepository.findById(id).orElseThrow(() -> new TeacherException("Admin is not found"));
+            adminRepository.delete(admin);
+        } else {
+            throw new AdminException("Impossible to delete last admin");
+        }
     }
 
     @Override
@@ -114,5 +124,29 @@ public class AdminServiceImpl implements AdminService {
         String query = "SELECT lesson.id, `begin`, `end`, `length`, group_level, teacher_id, first_name, second_name, last_name, status, `name` FROM lesson left outer join `group` on lesson.group_id = `group`.id left outer join dance d on `group`.dance_id = d.id left outer join person on lesson.teacher_id=person.id WHERE status='declined'";
         lessons.addAll(jdbcTemplate.query(query, new SheduleRowMapper()));
         return lessons;
+    }
+
+    public List<TeacherDto> getTeachersWithoutRole() {
+        List<TeacherDto> teachers = new ArrayList<>();
+        String query = "SELECT person.id, first_name, second_name, last_name, birthday, sex, e_mail, phone_number, name FROM person left outer join `role` on person.role_id = role.id WHERE `name` = 'ROLE_USER'";
+        teachers.addAll(jdbcTemplate.query(query, new Object[]{}, (resultSet, i) -> {
+            TeacherDto teacherDto = new TeacherDto();
+            teacherDto.setId(resultSet.getLong("id"));
+            teacherDto.setFirstName(resultSet.getString("first_name"));
+            teacherDto.setSecondName(resultSet.getString("second_name"));
+            teacherDto.setLastName(resultSet.getString("last_name"));
+            teacherDto.setSex(resultSet.getString("sex"));
+            teacherDto.setEmail(resultSet.getString("e_mail"));
+            teacherDto.setPhoneNumber(resultSet.getString("phone_number"));
+            teacherDto.setBirthday(resultSet.getDate("birthday"));
+            return teacherDto;
+        }));
+        return teachers;
+    }
+
+    public TeacherDto giveTeacherRole(Long teacherId) {
+        Teacher teacher = teacherRepository.getOne(teacherId);
+        teacher.setRole(roleRepository.findByName("ROLE_TEACHER"));
+        return conversionService.convert(teacherRepository.save(teacher), TeacherDto.class);
     }
 }
